match: nb_init -> 93.45% (SOURCE/netwin)
Reconstructs the whole netbios subsystem init: zero the session tables/buffers, probe via
is_netbios_avail, then set up 2 critical sections, 3 queue anchors (init_anchor), 9 manual-reset
events, and issue the NCBRESET + add-name sequence. Modeled the 14 file-static netbios globals in
_globals_model.h with real Win32 types (NCB/CRITICAL_SECTION/HANDLE from the real headers) — sizes
confirmed by address tiling; defined tag_Anchor (8 bytes, head/tail) reversed from init_anchor.
Signature is unsigned short(unsigned short,unsigned short) per the word stores. Residual is the
BaseAlloc/BaseFree __LINE__ arg (retail computes it via a const load; we pass __LINE__) — a
sanctioned deterministic hit. Overall %% dip is the expected windows.h cross-TU effect (max%% holds).

+++ include/SOURCE/netwin.h
@@ -4,7 +4,7 @@
 // symbols. Other TUs call them by including this header (no local externs).
 
 int is_netbios_avail(void);
-extern "C" int __fastcall nb_init(int, int);
+extern "C" unsigned short __fastcall nb_init(unsigned short, unsigned short);
 extern "C" int __fastcall nb_term(void);
 extern "C" int __fastcall nb_rcv(int, int);
 extern "C" int __fastcall nb_snd(int, int, int);
